power: qpnp-smbcharger: Hide charge counter
Some devices have a bug where querying the charge counter during deep
sleep takes an excessively long amount of time. BatteryService looks
up this prop during deep sleep, resulting in a very slow wakeup on these
devices.

Mido is affected with this bug. Add a filter to hide
the charge counter, so BatteryService does not attempt
to read the prop. It is unknown as to whether this
is a firmware or a hardware issue.

Change-Id: I4a4034090c165b795ddbbd1994b79afc61ce8127
Signed-off-by: leviathanaxeislit <scp.thedevil@gmail.com>